+++ scripts/lcoe_calculation.py
